fix: fallback for empty jsonform data
