+++ src/helpers/index.js
@@ -13,3 +13,4 @@ export { stringifyIfNeeded, parseIfNeeded } from "./transform";
 export { runAutomation } from "./automation";
 export { default as preventListItemClick } from "./prevent-list-item-click";
 export { default as createSession } from "./create-session";
+export { default as fetchJsonFormContent } from "./fetch-json-form-content";
